FINAL PUSH

+++ ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: jkrause <jkrause@student.42.us.org>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/04/10 11:22:00 by jkrause           #+#    #+#             */
-/*   Updated: 2017/06/20 00:32:39 by jkrause          ###   ########.fr       */
+/*   Updated: 2017/06/23 13:22:40 by jkrause          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 static int		trim_space(const char *str)
 {
-	int		i;
+	int			i;
 
 	i = 0;
 	while ((str[i] >= 9 && str[i] <= 13) || str[i] == 32)
@@ -24,9 +24,9 @@ static int		trim_space(const char *str)
 
 int				ft_atoi(const char *str)
 {
-	int		num;
-	int		i;
-	int		neg;
+	long long	num;
+	int			i;
+	int			neg;
 
 	num = 0;
 	neg = 0;
@@ -46,5 +46,5 @@ int				ft_atoi(const char *str)
 		if (str[i + 1] >= '0' && str[i + 1] <= '9')
 			num *= 10;
 	}
-	return (neg == 1 ? -num : num);
+	return (neg == 1 ? (int)-num : (int)num);
 }
